+++ CHANGES.md
@@ -1,5 +1,5 @@
 **Added:**
-* line number included for preprocess errors
+* preprocessor text replacing method (along with tests)
 
 **Changed:**
 * none